pyproject.toml
[tool.poetry]
name = "squad-calc-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.3"
sounddevice = "^0.5.1"
scipy = "^1.14.1"
openai = "^1.54.3"
keyboard = "^0.13.5"
python-dotenv = "^1.0.1"
pvporcupine = "^3.0.3"
webrtcvad = "^2.0.10"
setuptools = "^75.3.0"
torch = "^2.5.1"
silero-vad = "^5.1.2"
pygame = "^2.6.1"
websockets = "^13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
